backend/requirements.txt
black==23.9.1
Django==4.2.5
django-behaviors==0.5.1
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.4
flake8==6.1.0
gunicorn==21.2.0
isort==5.12.0
mixer==7.2.2
Pillow==10.0.1
psycopg2-binary==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
social-auth-app-django==5.3.0
